#94 使いまわしたいのでカスタムフック化した。
